Moved webcam code, added some test css
