+++ Frontend/insight/src/App.js
@@ -1,4 +1,4 @@
-import React, { useEffect, useRef } from "react";
+import React from "react";
 import Button from './components/Button';
 import MenuButton from './components/MenuButton';
 import Menu from './components/Menu';
@@ -14,36 +14,16 @@ import Card from './components/Card';
 
 function App() {
 
-  // Following code sourced from:
-  // https://itnext.io/accessing-the-webcam-with-javascript-and-react-33cbe92f49cb
-  const videoRef = useRef(null);
-
-  useEffect(() => {
-    getVideo();
-  }, [videoRef]);
-
-  const getVideo = () => {
-    navigator.mediaDevices
-      .getUserMedia({ video: { width: 400 } })
-      .then(stream => {
-        let video = videoRef.current;
-        video.srcObject = stream;
-        video.play();
-      })
-      .catch(err => {
-        console.error("error:", err);
-      });
-  };
-  // -----------------------------------------------
-
   return (
     <div className="App">
       <p>~~~~ Test Default Text ~~~~</p>
       <MenuButton />
       <HelpButton />
       <ReloadButton />
-      <video className="video-display" ref={videoRef} />
-      <ModeSwitcher />
+      <div className="webcam-block">
+        <VideoDisplay />
+        <ModeSwitcher />
+      </div>
       <CardStack />
       
     </div>
