+++ Frontend/insight/src/components/ModeSwitcher.js
@@ -3,8 +3,8 @@ import PropTypes from 'prop-types'
 
 function ModeSwitcher(props) {
     return (
-        <div>
-            
+        <div className="modeSwitcher">
+            <p>Example Placeholder Text</p>
         </div>
     )
 }
