+++ Frontend/insight/src/index.css
@@ -14,10 +14,26 @@ code {
     monospace;
 }
 
-.video-display {
+.webcam-block {
   position: absolute;
-  top: 32px;
-  left: 32px;
+  top: 16px;
+  left: 16px;
+  background-color: gray;
+  height: calc(100vh - 80px);
+  width: calc(100vw - 32px);
+}
+
+.video-display {
+  top: 0;
+  left: 0;
+  width: 100%;
   border-radius: 16px;
-  padding: 0;
+  background-color: red;
+}
+
+.mode-switcher {
+  display: inline-block;
+  border: 0px;
+  margin: 0px;
+  background-color: green;
 }
